DNM Checking CRC Cloud with image based on OCP 4.16.4
This is early test PR that uses new way for deploying the CRC -
crc-cloud [1] and image that is base on OCP 4.16.
with small modifications [2]. In the future, that image would be created
by the nodepool-builder, but right now it is just a PoC (Proof of concept).
Also is worth to mention, that the base job includes few workarounds
until the PR in the official projects are not merged [3][4][5]

[1] https://github.com/crc-org/crc-cloud
[2] https://softwarefactory-project.io/r/c/software-factory/sf-infra/+/31853
[3] https://github.com/crc-org/crc-cloud/pull/188
[4] https://github.com/crc-org/crc-cloud/pull/189
[5] https://github.com/crc-org/crc-cloud/pull/190